after login

+++ app/Http/Controllers/TasksController.php
@@ -8,20 +8,27 @@ use App\Task;
 
 class TasksController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    
     public function index()
     {
-        //$data = [];
-        /**if (\Auth::check()) { // 認証済みの場合
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
             // 認証済みユーザを取得
             $user = \Auth::user();
             // ユーザの投稿の一覧を作成日時の降順で取得
             $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
-           **/ 
+            
+            $data = [
+            'user' => $user,
+            'tasks' => $tasks,
+            
+            ];
+        }
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
+    }
+           
+           /**
        // メッセージ一覧を取得
         $tasks = Task::all();
        // }
@@ -31,36 +38,21 @@ class TasksController extends Controller
         ]);
         
     }
-    
-        /**
-        $data = [
-            'user' => $user,
-            'tasks' => $tasks,
-            ];
-        }
-        // Welcomeビューでそれらを表示
-        return view('welcome', $data);
-    }
     **/
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    
     public function create()
     {
         //
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
         $task = new Task;
         
         return view('tasks.create', ['task' => $task,]);
+        }
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+   
     public function store(Request $request)
     {
         
@@ -68,7 +60,8 @@ class TasksController extends Controller
             'content' => 'required|max:50',
             'status' => 'required|max:10',
             ]);
-        // メッセージを作成
+            
+    /**    // メッセージを作成
         $task = new Task;
         $task->content = $request->content;
         $task->status = $request->status;
@@ -76,15 +69,22 @@ class TasksController extends Controller
 
         // トップページへリダイレクトさせる
         return redirect('/');
+    **/
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
+
+        return redirect('/');
+        
     }
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+   
     public function show($id)
     {
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
 
@@ -92,16 +92,16 @@ class TasksController extends Controller
         return view('tasks.show', [
             'task' => $task,
         ]);
+        }
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+    
     public function edit($id)
     {
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
 
@@ -109,14 +109,9 @@ class TasksController extends Controller
         return view('tasks.edit', [
             'task' => $task,
         ]);
+        }
     }
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+    
     public function update(Request $request, $id)
     {
         $request->validate([
@@ -124,29 +119,41 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             ]);
         
-        
-        $task = Task::findOrFail($id);
-        
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-        
+            if (\Auth::check()) { // 認証済みの場合
+                // 認証済みユーザを取得
+                $user = \Auth::user();
+                    
+                $task = Task::findOrFail($id);
+                /** 
+                $task->content = $request->content;
+                $task->status = $request->status;
+                $task->save();
+                **/
+                // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+                $request->user()->tasks()->create([
+                    'content' => $request->content,
+                    'status' => $request->status,
+                ]);
+            }
         return redirect('/');
     }
     
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+    
     public function destroy($id)
     {
-        $task = Task::findOrFail($id);
-        
-        $task->delete();
-        
+            if (\Auth::check()) { // 認証済みの場合
+                        // 認証済みユーザを取得
+                        $user = \Auth::user();
+                $task = Task::findOrFail($id);
+                
+                $task->delete();
+                // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+                $request->user()->tasks()->create([
+                    'content' => $request->content,
+                    'status' => $request->status,
+                ]);
+            }
         return redirect('/');
     }
 }
